refactor: encapsulate FW rule processing
